Added Facebook authentication.
This was done with OAuth.io and Python Social Auth as after doing some
reading, this seemed like the easiest and most flexible approach.
Essentially, OAuth.io takes care of authentication on the client side
and the Facebook access token that is received is then send to PSA on
the server to perform double authentication. If this server side
authentication is successful, a Django REST API AuthToken is sent back
to the client and stored in a cookie called beLocalToken for later
authentication with the REST API.

We still need to determine how to categorize certain users as sellers
and others as customers, but I believe Django should allow us to do
this.