Added client side redirection for authentication
We are now returning a userType from the server that is either VEN (for
sellers) or CUS (for customers). Depending on which comes back, we can
redirect to the correct view on the client using the lines of code in
navbar.js that are currently commented out.

I also added a nice error message if the login process fails for some
reason. This is a good basic demonstration of AngularJS binding if you
want to figure out how to do it.

Finally, a StateService object was added to keep track of global
information such as user profile information.