Made some modifications to the data model.
Changed most DateFields to DateTimeFields to include time

Fixed a typo in the name ProductTag

Changed the addr_type field to an enum, as I think this is what we
want? Feel free to correct me if I misunderstood the purpose of this
field.

Removed the GeoLocation class, as we were essentially just mapping each
Geolocation to an address, so I figured we might as well just lump lat
and long straight into the address entity. This should also simplify
the creation of addresses on the client side.

Made Vendor inherit from django.contrib.auth.models.User so that it
inherits all the data fields that are already being stored in the
database.

Removed the stock attribute from Product, as I think it makes more
sense to have stock as an attribute on the SellerProductAtLocation
relation.