Added skeleton template for product cards
This required the addition of a few new scripts:

The first of these is holder.js, which is used for inserting
placeholder images into our application programatically. (As an aside,
a directive called holderFix was required to get this to play nicely
with AngularJS)

The second of these is angular-filter, which is currently used for
formatting currency. The rest of this commit is then relatively
straightforward. A few new serializers were created on the server side
to obtain the correct vendor information on the client, and the
endpoint created in the previous commit was hooked up in StateService
via getTrendingProducts().