Began database refactor
Changed all models to reflect new ER diagram. There are still several
broken views and serializers, but adding a location from the client
side is working with the new changes in place. Still to come are
view/serialization changes, as well as the ability to add repeating
dates from the UI.