Centered content
I’m not sure if this makes things better or worse, but I think what was
bothering me most was the fact that the midline of the content wasn’t
aligned with the midline of the container.

I also hid the mission statement for now since I think there were still
too many typefaces on the home tab. I’ll take another look at this
tomorrow with a fresh pair of eyes.