Updated tab appearance slightly.
Also disabled all active styles when a user is not on the main page.
This is what I’ve seen a few sites do, and to me, it makes more sense
than having (for instance) the Farmers tab highlighted when viewing the
Vendor Details page.

I also disabled auto-scrolling to anchors (#’s) in app.js to prevent
the weird scrolling behaviour we were getting when switching tabs.