Created a (somewhat) hacky directive to fix maps
I’m not sure if this is something we actually want to push to master
since it involves listening to multiple scroll events with jQuery, but
this is the only way that I could get this to work since Bootstrap’s
Affix (and all other related plugins that use position:fixed) don’t
work with fluid column layouts. As a result, I’m setting the top CSS
property on every scroll event in order to mimic the appearance of
position: fixed.

To use this directive, use ‘affix’ and specify an offset-top attribute
that is used to determine when any element with the ‘affix’ attribute
should stop scrolling with the rest of the page and remain in a fixed
position.