Basic UI and logic for recurring selling locations
After much thought, I decided to move the recurrence objects into
SellerLocations and Markets respectively, as having one for every
single hour in an OpeningHours object introduced a HUGE amount of
overhead for no reason.

Now, we’ll be able to quickly determine if a SellerLocation or Market
is recurring without having to introduce a bunch of changes to the
hours logic that we had going previously. There’s still a lot of work
that needs to be done here, but I think this is a decent start.