Modified recurrences slightly
The start date of recurrence objects in the database now contains the
Monday (start of the week) of the specified start date. Then, when we
return recurrences from the database, the dates returned will be the
Mondays (starts of the week) for each recurrence. Arithmetic can then
be performed on the client side to display the correct dates in
whichever fashion we see fit.