Updated UI strings
We are now displaying “Closed for the season” in the event that hours
for markets are null, or if the recurrence date range for a given
market is in the past.

We are displaying “Hours for” either “this week” or “<futureWeekRange>”
in the event that we are either within the recurrence range, or the
recurrence range is upcoming respectively.